Struct msgpacknet::CloseGuard

pub struct CloseGuard<M: Message, N: NodeId, I: InitMessage>(_);

A guard that closes the node when dropped.

Implementation details require the node to exist in multiple reference counted copies. Therefore this wrapper is needed to implement the drop behavior. Except for this drop behavior this struct can be used like the node struct that it encapsulates.