Trait msgpacknet::InitMessage

pub trait InitMessage: Serialize + Deserialize + Send + Sync + Clone + Debug + 'static { }

The trait used for initialization messages

This is the type of message that will be exchanged during the initialization phase. It could be used to differenciate between different protocols, versions and capabilities and to decide whether to accept a connection or not. This trait is implemented automatically for all types that fulfill the requirements.